Move hook back to folder local to it's usage
