+++ packages/harpguru-core/src/components/harp-face/harp-face.tsx
@@ -2,9 +2,8 @@ import { View } from 'react-native'
 import React from 'react'
 
 import { HarpFaceFragment } from '../harp-face-fragment'
-import { useOctaveColumnGroups } from '../../hooks'
 
-import { useStyles } from './hooks'
+import { useStyles, useOctaveColumnGroups } from './hooks'
 
 type HarpFaceProps = {
   readonly harpfaceIndex: 'harpface1' | 'harpface2'
