+++ packages/harpguru-core/src/components/harp-face/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useStyles } from './use-styles'
+export { useOctaveColumnGroups } from './use-octave-column-groups'
