+++ packages/harpguru-core/src/components/harp-face/hooks/use-octave-column-groups/array-has-root.test.ts
