+++ packages/harpguru-core/src/components/harp-face/hooks/use-octave-column-groups/use-octave-column-groups.test.ts
@@ -1,6 +1,6 @@
 import { useGlobal } from 'reactn'
 
-import { buildMockUseGlobalImplementation } from '../../test-resources'
+import { buildMockUseGlobalImplementation } from '../../../../test-resources'
 
 import { useOctaveColumnGroups } from './use-octave-column-groups'
