+++ packages/harpguru-core/src/components/harp-face/hooks/use-octave-column-groups/use-octave-column-groups.ts
@@ -1,8 +1,8 @@
 import { useGlobal } from 'reactn'
 import type { Degree, HarpFaceMatrix } from 'harpparts'
 
-import { extractHarpFaceFacts } from '../../utils'
-import { transposeMatrix } from '../../packages/transpose-matrix'
+import { extractHarpFaceFacts } from '../../../../utils'
+import { transposeMatrix } from '../../../../packages/transpose-matrix'
 
 import { getOctaveColumnGroups } from './get-octave-column-groups'
 import type { ColumnRanges } from './get-octave-column-groups'
